Fix numeric type to Long for ThreadPool information bean (cherry picked from commit 256fac4)

+++ argos-agent/src/main/scala/io/argos/agent/bean/MetricsAggregators.scala
@@ -3,7 +3,7 @@ package io.argos.agent.bean
 /**
  * This class groups all independent values relative to a type of ThreadPool (MutationStage, ReadRepairStage...)
  */
-case class ThreadPoolStats(`type`: String, activeTasks: Int, completedTasks: Int, currentBlockedTasks: Long, maxPoolSize: Int, pendingTasks: Int, totalBlockedTasks: Long)
+case class ThreadPoolStats(`type`: String, activeTasks: Long, completedTasks: Long, currentBlockedTasks: Long, maxPoolSize: Long, pendingTasks: Long, totalBlockedTasks: Long)
 
 /**
  * This class contains all rates of dropped messages for a given type of request (rate unit = events/second)
